stylizedTimestamp is a better name
